requirements/base.txt
django==3.1.4
djangorestframework==3.12.2
django-environ==0.4.5
mysqlclient==2.0.2
